Fixing findbugs errors

+++ hbase_pipeline/src/main/java/com/splicemachine/derby/hbase/CountingPipelineMeter.java
@@ -27,7 +27,7 @@ public class CountingPipelineMeter implements PipelineMeter{
 
     @Override
     public double throughput(){
-        return successCounter.get()/(System.nanoTime()-startupTimestamp);
+        return ((double)successCounter.get())/(System.nanoTime()-startupTimestamp);
     }
 
     @Override
